Factor CXXQuery/CXXBuilder out of CXXInterface
This allows us to grow the three pieces independently:
 * Core interface: Knows how to build the compile units, parse files, etc.
 * Query interafce: Knows all about AST types and functions.
 * Builder interface: Know how to build new AST nodes.

The plan is for CXXInterface to have both query and builder, and pass
those to the users that only need one or the other.

The Verona type checker only needs to query types and function names,
no need to build new AST nodes on the C++ side. But type inference will
need to construct C++ functions for specialised types (including
constructor).

The two interfaces will grow substantially and it's better to keep them
separate for now, to avoid mixing const checks with non-const builder
logic at an early design stage. It also helps keeping the files as short
as possible.
