+++ src/interop/CXXInterface.h
@@ -4,11 +4,11 @@
 #pragma once
 
 #include "CXXType.h"
+#include "CXXQuery.h"
+#include "CXXBuilder.h"
 #include "Compiler.h"
 #include "FS.h"
 
-#include <clang/ASTMatchers/ASTMatchFinder.h>
-#include <clang/ASTMatchers/ASTMatchers.h>
 #include <clang/CodeGen/ModuleBuilder.h>
 #include <clang/Frontend/CompilerInstance.h>
 #include <clang/Frontend/CompilerInvocation.h>
@@ -20,9 +20,6 @@
 #include <clang/Lex/HeaderSearchOptions.h>
 #include <clang/Lex/Preprocessor.h>
 #include <clang/Lex/PreprocessorOptions.h>
-#include <clang/Sema/Sema.h>
-#include <clang/Sema/Template.h>
-#include <clang/Sema/TemplateDeduction.h>
 #include <clang/Serialization/ASTWriter.h>
 #include <clang/Tooling/Tooling.h>
 #include <llvm/IR/LLVMContext.h>
@@ -55,6 +52,10 @@ namespace verona::interop
     std::unique_ptr<Compiler> Clang;
     /// Virtual file system (compiler unit, pch, headers)
     FileSystem FS;
+    /// Query system
+    std::unique_ptr<CXXQuery> query;
+    /// Build system
+    std::unique_ptr<CXXBuilder> builder;
 
     /**
      * Creates new AST consumers to add the AST back into the interface.
@@ -90,37 +91,6 @@ namespace verona::interop
         return std::make_unique<Collector>(interface);
       }
     } factory;
-
-    /**
-     * Simple handler for indirect dispatch on a Clang AST matcher.
-     *
-     * Use:
-     * ```
-     *  void myfunc(MatchFinder::MatchResult &);
-     *  MatchFinder f;
-     *  f.addMatcher(new HandleMatch(myfunc));
-     *  f.matchAST(*ast);
-     *  // If matches, runs `myfunc` on the matched AST node.
-     * ```
-     */
-    class HandleMatch : public clang::ast_matchers::MatchFinder::MatchCallback
-    {
-      std::function<void(
-        const clang::ast_matchers::MatchFinder::MatchResult& Result)>
-        handler;
-      void
-      run(const clang::ast_matchers::MatchFinder::MatchResult& Result) override
-      {
-        handler(Result);
-      }
-
-    public:
-      HandleMatch(
-        std::function<
-          void(const clang::ast_matchers::MatchFinder::MatchResult& Result)> h)
-      : handler(h)
-      {}
-    };
 
     /**
      * Pre-compiled header action, to create the PCH consumers for PCH
@@ -198,45 +168,6 @@ namespace verona::interop
         new llvm::SmallVectorMemoryBuffer(std::move(pchOutBuffer)));
     }
 
-    /// Maps between CXXType and Clang's types.
-    clang::QualType typeForBuiltin(CXXType::BuiltinTypeKinds ty)
-    {
-      switch (ty)
-      {
-        case CXXType::BuiltinTypeKinds::Float:
-          return ast->FloatTy;
-        case CXXType::BuiltinTypeKinds::Double:
-          return ast->DoubleTy;
-        case CXXType::BuiltinTypeKinds::Bool:
-          return ast->BoolTy;
-        case CXXType::BuiltinTypeKinds::SChar:
-          return ast->SignedCharTy;
-        case CXXType::BuiltinTypeKinds::Char:
-          return ast->CharTy;
-        case CXXType::BuiltinTypeKinds::UChar:
-          return ast->UnsignedCharTy;
-        case CXXType::BuiltinTypeKinds::Short:
-          return ast->ShortTy;
-        case CXXType::BuiltinTypeKinds::UShort:
-          return ast->UnsignedShortTy;
-        case CXXType::BuiltinTypeKinds::Int:
-          return ast->IntTy;
-        case CXXType::BuiltinTypeKinds::UInt:
-          return ast->UnsignedIntTy;
-        case CXXType::BuiltinTypeKinds::Long:
-          return ast->LongTy;
-        case CXXType::BuiltinTypeKinds::ULong:
-          return ast->UnsignedLongTy;
-        case CXXType::BuiltinTypeKinds::LongLong:
-          return ast->LongLongTy;
-        case CXXType::BuiltinTypeKinds::ULongLong:
-          return ast->UnsignedLongLongTy;
-      }
-      // TODO: This is wrong but silences a warning, need to know what's the
-      // correct behaviour here.
-      return ast->VoidTy;
-    }
-
   public:
     /**
      * CXXInterface c-tor. Creates the internal compile unit, include the
@@ -279,364 +210,10 @@ namespace verona::interop
       // Executing the action consumes the AST.  Reset the compiler instance to
       // refer to the AST that it just parsed and create a Sema instance.
       Clang->setASTMachinery(factory.newASTConsumer(), ast);
-    }
 
-    /**
-     * Gets an {class | template | enum} type from the source AST by name.
-     * The name must exist and be fully qualified and it will match in the
-     * order specified above.
-     *
-     * We don't need to find builtin types because they're pre-defined in the
-     * language and represented in CXXType directly.
-     *
-     * TODO: Change this method to receive a list of names and return a list
-     * of types (or some variation over mutliple types at the same time).
-     */
-    CXXType getType(std::string name)
-    {
-      name = "::" + name;
-      clang::ast_matchers::MatchFinder finder;
-      const clang::EnumDecl* foundEnum = nullptr;
-      const clang::CXXRecordDecl* foundClass = nullptr;
-      const clang::ClassTemplateDecl* foundTemplateClass = nullptr;
-
-      finder.addMatcher(
-        clang::ast_matchers::cxxRecordDecl(clang::ast_matchers::hasName(name))
-          .bind("id"),
-        new HandleMatch(
-          [&](const clang::ast_matchers::MatchFinder::MatchResult& Result) {
-            auto* decl = Result.Nodes.getNodeAs<clang::CXXRecordDecl>("id")
-                           ->getDefinition();
-            if (decl)
-            {
-              foundClass = decl;
-            }
-          }));
-      finder.addMatcher(
-        clang::ast_matchers::classTemplateDecl(
-          clang::ast_matchers::hasName(name))
-          .bind("id"),
-        new HandleMatch(
-          [&](const clang::ast_matchers::MatchFinder::MatchResult& Result) {
-            auto* decl = Result.Nodes.getNodeAs<clang::ClassTemplateDecl>("id");
-            if (decl)
-            {
-              foundTemplateClass = decl;
-            }
-          }));
-      finder.addMatcher(
-        clang::ast_matchers::enumDecl(clang::ast_matchers::hasName(name))
-          .bind("id"),
-        new HandleMatch(
-          [&](const clang::ast_matchers::MatchFinder::MatchResult& Result) {
-            auto* decl = Result.Nodes.getNodeAs<clang::EnumDecl>("id");
-            if (decl)
-            {
-              foundEnum = decl;
-            }
-          }));
-      finder.matchAST(*ast);
-
-      // Should onlyt match one, so this is fine.
-      if (foundTemplateClass)
-      {
-        return CXXType(foundTemplateClass);
-      }
-      if (foundClass)
-      {
-        return CXXType(foundClass);
-      }
-      if (foundEnum)
-      {
-        return CXXType(foundEnum);
-      }
-
-      // If didn't match any type, check for builtins
-      return llvm::StringSwitch<CXXType>(name)
-        .Case("::bool", CXXType::getBoolean())
-        .Case("::unsigned char", CXXType::getUnsignedChar())
-        .Case("::char", CXXType::getChar())
-        .Case("::signed char", CXXType::getSignedChar())
-        .Case("::short", CXXType::getShort())
-        .Case("::unsigned short", CXXType::getUnsignedShort())
-        .Case("::int", CXXType::getInt())
-        .Case("::unsigned int", CXXType::getUnsignedInt())
-        .Case("::long", CXXType::getLong())
-        .Case("::unsigned long", CXXType::getUnsignedLong())
-        .Case("::long long", CXXType::getLongLong())
-        .Case("::unsigned long long", CXXType::getUnsignedLongLong())
-        .Case("::float", CXXType::getFloat())
-        .Case("::double", CXXType::getDouble())
-        // Otherwise, just return empty invalid type
-        .Default(CXXType());
-    }
-
-    /// Return the size in bytes of the specified type.
-    uint64_t getTypeSize(CXXType& t)
-    {
-      assert(t.kind != CXXType::Kind::Invalid);
-      if (t.sizeAndAlign.Width == 0)
-      {
-        clang::QualType ty = getQualType(t);
-        t.sizeAndAlign = ast->getTypeInfo(ty);
-      }
-      return t.sizeAndAlign.Width / 8;
-    }
-
-    /// Return the qualified type for a CXXType
-    /// FIXME: Do we really need to expose this?
-    clang::QualType getQualType(CXXType ty)
-    {
-      switch (ty.kind)
-      {
-        case CXXType::Kind::Invalid:
-        case CXXType::Kind::TemplateClass:
-          // TODO: Fix template class
-          return clang::QualType{};
-        case CXXType::Kind::SpecializedTemplateClass:
-        case CXXType::Kind::Class:
-          return ast->getRecordType(ty.getAs<clang::CXXRecordDecl>());
-        case CXXType::Kind::Enum:
-          return ast->getEnumType(ty.getAs<clang::EnumDecl>());
-        case CXXType::Kind::Builtin:
-          return typeForBuiltin(ty.builtTypeKind);
-      }
-      // TODO: This is wrong but silences a warning, need to know what's the
-      // correct behaviour here.
-      return ast->VoidTy;
-    }
-
-    /// Returns the type as a template argument.
-    clang::TemplateArgument createTemplateArgumentForType(CXXType t)
-    {
-      switch (t.kind)
-      {
-        case CXXType::Kind::Invalid:
-        case CXXType::Kind::TemplateClass:
-          return clang::TemplateArgument{};
-        case CXXType::Kind::SpecializedTemplateClass:
-        case CXXType::Kind::Class:
-          return clang::TemplateArgument{
-            ast->getRecordType(t.getAs<clang::CXXRecordDecl>())};
-        case CXXType::Kind::Enum:
-          return clang::TemplateArgument{
-            ast->getEnumType(t.getAs<clang::EnumDecl>())};
-        case CXXType::Kind::Builtin:
-          return clang::TemplateArgument{typeForBuiltin(t.builtTypeKind)};
-      }
-      return nullptr;
-    }
-
-    /// Returns the integral literal as a template value.
-    /// TODO: C++20 accepts floating point too
-    clang::TemplateArgument createTemplateArgumentForIntegerValue(
-      CXXType::BuiltinTypeKinds ty, uint64_t value)
-    {
-      assert(CXXType::isIntegral(ty));
-      clang::QualType qualTy = typeForBuiltin(ty);
-      auto info = ast->getTypeInfo(qualTy);
-      llvm::APInt val{static_cast<unsigned int>(info.Width), value};
-      auto* literal = clang::IntegerLiteral::Create(
-        *ast, val, qualTy, clang::SourceLocation{});
-      return clang::TemplateArgument(literal);
-    }
-
-    /**
-     * Instantiate the class template specialisation at the end of the main
-     * file, if not yet done.
-     */
-    CXXType instantiateClassTemplate(
-      CXXType& classTemplate, llvm::ArrayRef<clang::TemplateArgument> args)
-    {
-      if (classTemplate.kind != CXXType::Kind::TemplateClass)
-      {
-        return CXXType{};
-      }
-
-      auto& S = Clang->getSema();
-
-      // Check if this specialisation is already present in the AST
-      // (declaration, definition, used).
-      clang::ClassTemplateDecl* ClassTemplate =
-        classTemplate.getAs<clang::ClassTemplateDecl>();
-      void* InsertPos = nullptr;
-      clang::ClassTemplateSpecializationDecl* Decl =
-        ClassTemplate->findSpecialization(args, InsertPos);
-      if (!Decl)
-      {
-        // This is the first time we have referenced this class template
-        // specialization. Create the canonical declaration and add it to
-        // the set of specializations.
-        Decl = clang::ClassTemplateSpecializationDecl::Create(
-          *ast,
-          ClassTemplate->getTemplatedDecl()->getTagKind(),
-          ClassTemplate->getDeclContext(),
-          ClassTemplate->getTemplatedDecl()->getBeginLoc(),
-          ClassTemplate->getLocation(),
-          ClassTemplate,
-          args,
-          nullptr);
-        ClassTemplate->AddSpecialization(Decl, InsertPos);
-      }
-      // If specialisation hasn't been directly declared yet (by the user),
-      // instantiate the declaration.
-      if (Decl->getSpecializationKind() == clang::TSK_Undeclared)
-      {
-        clang::MultiLevelTemplateArgumentList TemplateArgLists;
-        TemplateArgLists.addOuterTemplateArguments(args);
-        S.InstantiateAttrsForDecl(
-          TemplateArgLists, ClassTemplate->getTemplatedDecl(), Decl);
-      }
-      // If specialisation hasn't been defined yet, create its definition at the
-      // end of the file.
-      clang::ClassTemplateSpecializationDecl* Def =
-        clang::cast_or_null<clang::ClassTemplateSpecializationDecl>(
-          Decl->getDefinition());
-      if (!Def)
-      {
-        clang::SourceLocation InstantiationLoc = Clang->getEndOfFileLocation();
-        assert(InstantiationLoc.isValid());
-        S.InstantiateClassTemplateSpecialization(
-          InstantiationLoc, Decl, clang::TSK_ExplicitInstantiationDefinition);
-        Def = clang::cast<clang::ClassTemplateSpecializationDecl>(
-          Decl->getDefinition());
-      }
-      return CXXType{Def};
-    }
-
-    /**
-     * Get the template especialization with args.
-     */
-    clang::QualType getTemplateSpecializationType(
-      const clang::NamedDecl* decl,
-      llvm::ArrayRef<clang::TemplateArgument> args)
-    {
-      clang::TemplateName templ{clang::dyn_cast<clang::TemplateDecl>(
-        const_cast<clang::NamedDecl*>(decl))};
-      return ast->getTemplateSpecializationType(templ, args);
-    }
-
-    /**
-     * Get the canonical template especialization with args.
-     */
-    clang::QualType getCanonicalTemplateSpecializationType(
-      const clang::NamedDecl* decl,
-      llvm::ArrayRef<clang::TemplateArgument> args)
-    {
-      clang::TemplateName templ{clang::dyn_cast<clang::TemplateDecl>(
-        const_cast<clang::NamedDecl*>(decl))};
-      return ast->getCanonicalTemplateSpecializationType(templ, args);
-    }
-
-    /**
-     * Instantiate a new function at the end of the main file, if not yet done.
-     */
-    clang::FunctionDecl* instantiateFunction(
-      const char* name, llvm::ArrayRef<CXXType> args, CXXType ret)
-    {
-      auto* DC = ast->getTranslationUnitDecl();
-      clang::SourceLocation loc = Clang->getEndOfFileLocation();
-      clang::IdentifierInfo& fnNameIdent = ast->Idents.get(name);
-      clang::DeclarationName fnName{&fnNameIdent};
-      clang::FunctionProtoType::ExtProtoInfo EPI;
-
-      // Get type of args/ret, function
-      llvm::SmallVector<clang::QualType> argTys;
-      for (auto argTy : args)
-        argTys.push_back(getQualType(argTy));
-      auto retTy = getQualType(ret);
-      clang::QualType fnTy = ast->getFunctionType(retTy, argTys, EPI);
-
-      // Create a new function
-      auto func = clang::FunctionDecl::Create(
-        *ast,
-        DC,
-        loc,
-        loc,
-        fnName,
-        fnTy,
-        ast->getTrivialTypeSourceInfo(fnTy),
-        clang::StorageClass::SC_None);
-
-      // Define all arguments
-      llvm::SmallVector<clang::ParmVarDecl*, 4> argDecls;
-      size_t argID = 0;
-      for (auto argTy : argTys)
-      {
-        std::string argName = "arg" + std::to_string(argID++);
-        clang::IdentifierInfo& ident = ast->Idents.get(argName);
-        clang::ParmVarDecl* arg = clang::ParmVarDecl::Create(
-          *ast,
-          func,
-          loc,
-          loc,
-          &ident,
-          argTy,
-          nullptr,
-          clang::StorageClass::SC_None,
-          nullptr);
-        argDecls.push_back(arg);
-      }
-
-      // Set function argument list
-      func->setParams(argDecls);
-
-      // Associate with the translation unit
-      func->setLexicalDeclContext(DC);
-      DC->addDecl(func);
-
-      return func;
-    }
-
-    /**
-     * Create a function argument
-     *
-     * FIXME: Do we want to have this as part of instantiateFunction?
-     */
-    clang::ParmVarDecl* createFunctionArgument(
-      const char* name, CXXType& ty, clang::FunctionDecl* func)
-    {
-      clang::SourceLocation loc = func->getLocation();
-      clang::IdentifierInfo& ident = ast->Idents.get(name);
-      clang::ParmVarDecl* arg = clang::ParmVarDecl::Create(
-        *ast,
-        func,
-        loc,
-        loc,
-        &ident,
-        getQualType(ty),
-        nullptr,
-        clang::StorageClass::SC_None,
-        nullptr);
-      func->setParams({arg});
-      return arg;
-    }
-
-    /**
-     * Create integer constant literal
-     *
-     * TODO: Create all the ones we use in the template specialisation
-     */
-    clang::IntegerLiteral*
-    createIntegerLiteral(unsigned int len, unsigned long val)
-    {
-      llvm::APInt num{len, val};
-      auto* lit = clang::IntegerLiteral::Create(
-        *ast, num, getQualType(CXXType::getInt()), clang::SourceLocation{});
-      return lit;
-    }
-
-    /**
-     * Create a return instruction
-     *
-     * TODO: Create all the ones we use in code generation
-     */
-    clang::ReturnStmt* createReturn(clang::Expr* val, clang::FunctionDecl* func)
-    {
-      auto retStmt =
-        clang::ReturnStmt::Create(*ast, func->getLocation(), val, nullptr);
-      func->setBody(retStmt);
-      return retStmt;
+      // Setup query/build system from AST
+      query = std::make_unique<CXXQuery>(ast, Clang.get());
+      builder = std::make_unique<CXXBuilder>(ast, Clang.get(), query.get());
     }
 
     /**
@@ -662,6 +239,18 @@ namespace verona::interop
     const Compiler* getCompiler() const
     {
       return Clang.get();
+    }
+
+    /// Get query system
+    const CXXQuery* getQuery() const
+    {
+      return query.get();
+    }
+
+    /// Get builder system
+    const CXXBuilder* getBuilder() const
+    {
+      return builder.get();
     }
   };
 } // namespace verona::interop
